Fix type circular dependencies
