+++ server/data/orchestrationApiTypes.ts
@@ -37,11 +37,3 @@ export type PrisonRegisterPrisonDto = components['schemas']['PrisonRegisterPriso
 export type VisitDto = components['schemas']['VisitDto']
 
 export type VisitorInfoDto = components['schemas']['VisitorInfoDto']
-
-export interface Visitor extends VisitorInfoDto {
-  visitorDisplayId: string
-  adult: boolean
-  eligible?: boolean
-  banned?: boolean
-  banExpiryDate?: string
-}
