+++ server/services/bookerService.test.ts
@@ -1,8 +1,8 @@
 import { BadRequest, TooManyRequests } from 'http-errors'
-import BookerService, { Prisoner, VisitorsByStatus } from './bookerService'
+import BookerService, { Prisoner, Visitor, VisitorsByStatus } from './bookerService'
 import TestData from '../routes/testutils/testData'
 import { createMockHmppsAuthClient, createMockOrchestrationApiClient } from '../data/testutils/mocks'
-import { BookerPrisonerValidationErrorResponse, Visitor } from '../data/orchestrationApiTypes'
+import { BookerPrisonerValidationErrorResponse } from '../data/orchestrationApiTypes'
 import { SanitisedError } from '../sanitisedError'
 import logger from '../../logger'
 import { createMockRateLimitService } from './testutils/mocks'
